Compensatory changes for AEGeAn updates [skip ci]
There has been a recent flurry of AEGeAn development (currently on the `refine` branch), and this commit brings HymHub up-to-date with the latest AEGeAn updates.
The most important of these is better handling of overlapping gene models, and an improved way of accounting for genome composition.
These are manifested in different GFF3 attributes that need to be handled differently, thus this commit.