Transcript IDs to accessions, many changes (closes #11)
Previously, HymHub used the ID attribute of mRNAs and other transcript types to refer to transcripts and resolve their relationships with other features like iLoci, exons, and introns.
While the ID attribute is guaranteed to be unique within a particular file, it is not necessarily unique or even consistent across multiple files.
This has gone unnoticed for quite a while, since very little has been done by way of analysis of exons and introns other than aggregate statistics on the whole set.
Recently when trying to pull out subsets of, for example, exons from conserved genes, the exons pulled out didn't seem to match the mRNAs requested.
This commit introduces many changes system-wide, and requires compensatory changes that have been committed to the AEGeAn `refine` branch, but ultimately hopes to address this issue.